+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Count markdown TODOs as "TODO" when showing totals in the status bar
 - Use a default colour for ruler highlights
 - Add better explanation of useBuiltInExcludes setting
+- Add support for compact folders in tree
 
 ## v0.0.166 - 2020-01-22